requirements.txt
Pillow
discord.py[voice]